+++ packages/astralis-ui/src/components/disclosure/collapsible/index.ts
@@ -1,0 +1,14 @@
+import { CollapsibleRoot } from "./components/collapsible-root";
+import { CollapsibleTrigger, CollapsibleContent } from "./components/collapsible-parts";
+
+export const Collapsible = Object.assign(CollapsibleRoot, {
+  Trigger: CollapsibleTrigger,
+  Content: CollapsibleContent,
+});
+
+export { CollapsibleRoot, CollapsibleTrigger, CollapsibleContent };
+export type {
+  CollapsibleProps,
+  CollapsibleTriggerProps,
+  CollapsibleContentProps,
+} from "./collapsible.types";