+++ packages/astralis-ui/src/components/overlay/index.ts
@@ -1,5 +1,6 @@
 export * from "./modal";
 export * from "./drawer";
 export * from "./popover";
+export * from "./popout";
 export * from "./tooltip";
 export { Portal } from "./portal";